swagger settings with dependencies added
