+++ pom.xml
@@ -92,14 +92,23 @@
     </dependency>
 
     <dependency>
-      <groupId>org.hibernate.common</groupId>
-      <artifactId>hibernate-commons-annotations</artifactId>
-      <version>5.0.4.Final</version>
+      <groupId>io.springfox</groupId>
+      <artifactId>springfox-swagger2</artifactId>
+      <version>3.0.0</version>
     </dependency>
 
     <dependency>
-      <groupId></groupId>
-      <artifactId></artifactId>
+      <groupId>io.springfox</groupId>
+      <artifactId>springfox-swagger-ui</artifactId>
+      <version>3.0.0</version>
+    </dependency>
+
+
+    <!-- https://mvnrepository.com/artifact/com.fasterxml.jackson.core/jackson-databind -->
+    <dependency>
+      <groupId>com.fasterxml.jackson.core</groupId>
+      <artifactId>jackson-databind</artifactId>
+      <version>2.14.2</version>
     </dependency>
 
     <dependency>
@@ -107,7 +116,6 @@
       <artifactId>springdoc-openapi-ui</artifactId>
       <version>1.6.9</version>
     </dependency>
-
   </dependencies>
 
   <build>
