+++ src/main/java/com/customerService/intsv/controller/TestController.java
@@ -1,0 +1,15 @@
+package com.customerService.intsv.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class TestController {
+
+    @GetMapping("/test")
+    public String test(){
+        return "Stupid Swagger!!!";
+    }
+
+}
