+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 #dataSource
 spring.datasource.url=jdbc:postgresql://localhost/CustomerService
 spring.datasource.username=postgres
-spring.datasource.password=admin
+spring.datasource.password=postgres
 spring.datasource.driver-class-name=org.postgresql.Driver
 
 #jpa
